feat: update README with detailed features and architecture, enhance user routes with authentication, and improve user service caching logic
