+++ src/modules/users/infrastructure/database/repositories/UserRepository.ts
@@ -4,14 +4,9 @@ import { IRegisterUser } from "../../../domain/models/IRegisterUser";
 import { IUser } from "../../../domain/models/IUser";
 import { IUserRepository } from "../../../domain/repositories/IUserRepository";
 import User from "../models/user.model";
-import { container } from "tsyringe";
-import RedisService from "../../../../../infrastructure/redis/Services/Redis.Service";
-
 class UserRepository implements IUserRepository {
   async delete(id: string): Promise<IUser | null> {
     const user: IUser | null = await User.findByIdAndDelete(id);
-    const redis = container.resolve(RedisService);
-    await redis.delete("users", "*");
     return user;
   }
   async findOne(id: string): Promise<IUser | null> {
@@ -19,19 +14,12 @@ class UserRepository implements IUserRepository {
     return user;
   }
   async find(data: IFindUsers): Promise<IUser[]> {
-    const redis = container.resolve(RedisService);
-    const cachedData = await redis.get("users", data);
-    if (cachedData) {
-      const dataObj = JSON.parse(cachedData);
-      return dataObj;
-    }
     const query = new ApiFeatures(User.find({}), data)
       .filter()
       .limit()
       .paginate()
       .sort();
     const users = await query.exec();
-    await redis.set("users", data, users);
     return users;
   }
   async findByEmail(email: string): Promise<IUser | null> {
@@ -40,8 +28,6 @@ class UserRepository implements IUserRepository {
   }
   async register({ name, email, password }: IRegisterUser): Promise<IUser> {
     const user: IUser = await User.create({ name, email, password });
-    const redis = container.resolve(RedisService);
-    await redis.delete("users", "*");
     return user;
   }
 }
